[FEAT] 5일 회의 코드 정리
