--- code/requirements.txt
@@ -1,5 +1,0 @@
-opencv-python-headless==4.10.0.84
-pandas==2.2.3
-scikit-learn==1.5.2
-albumentations==1.4.18
-matplotlib==3.9.2
