+++ requirements.txt
@@ -1,0 +1,199 @@
+expecttest==0.1.6
+triton==2.1.0
+mkl-fft==1.3.8
+torchaudio==2.1.0
+wcwidth==0.2.5
+beautifulsoup4==4.12.2
+setuptools==68.0.0
+libarchive-c==2.9
+psutil==5.9.0
+chardet==4.0.0
+six==1.16.0
+truststore==0.8.0
+conda_index==0.3.0
+pycosat==0.6.4
+filelock==3.9.0
+jsonpatch==1.32
+torchelastic==0.2.2
+networkx==3.1
+urllib3==1.26.16
+jedi==0.18.1
+pip==23.2.1
+Pillow==9.4.0
+zstandard==0.19.0
+tomli==2.0.1
+conda-libmamba-solver==23.7.0
+pycparser==2.21
+soupsieve==2.5
+pure-eval==0.2.2
+ruamel.yaml==0.17.21
+pluggy==1.0.0
+packaging==23.1
+pyOpenSSL==23.2.0
+MarkupSafe==2.1.1
+decorator==5.1.1
+conda-content-trust==0.2.0
+python-etcd==0.4.5
+types-dataclasses==0.6.6
+fsspec==2023.9.2
+torchvision==0.16.0
+gmpy2==2.1.2
+traitlets==5.7.1
+cffi==1.15.1
+requests==2.31.0
+click==8.0.4
+conda-build==3.27.0
+attrs==23.1.0
+typing_extensions==4.7.1
+typing_extensions==4.8.0
+prompt-toolkit==3.0.36
+cryptography==41.0.3
+jsonpointer==2.1
+conda_package_streaming==0.9.0
+ruamel.yaml.clib==0.2.6
+sympy==1.11.1
+sympy==1.12
+brotlipy==0.7.0
+exceptiongroup==1.0.4
+ptyprocess==0.7.0
+dnspython==2.4.2
+parso==0.8.3
+astunparse==1.6.3
+mkl-random==1.2.4
+pkginfo==1.9.6
+pickleshare==0.7.5
+sortedcontainers==2.4.0
+executing==0.8.3
+more-itertools==8.12.0
+Pygments==2.15.1
+charset-normalizer==2.0.4
+PyYAML==6.0
+tqdm==4.65.0
+mkl-service==2.4.0
+PySocks==1.7.1
+asttokens==2.0.5
+stack-data==0.2.0
+idna==3.4
+conda-package-handling==2.2.0
+toolz==0.12.0
+torch==2.1.0
+ipython==8.15.0
+wheel==0.41.2
+pytz==2023.3.post1
+conda==23.9.0
+mpmath==1.3.0
+pexpect==4.8.0
+hypothesis==6.87.1
+boltons==23.0.0
+matplotlib-inline==0.1.6
+numpy==1.26.0
+Jinja2==3.1.2
+backcall==0.2.0
+libmambapy==1.4.1
+platformdirs==4.3.6
+optree==0.13.1
+optuna==4.1.0
+pretrainedmodels==0.7.4
+pyzmq==25.1.2
+ml-dtypes==0.4.1
+greenlet==3.1.1
+gitdb==4.0.11
+mistune==3.0.2
+pyparsing==3.2.0
+setproctitle==1.3.3
+SQLAlchemy==2.0.36
+albucore==0.0.17
+wandb==0.18.6
+Markdown==3.7
+GitPython==3.1.43
+lion-pytorch==0.2.2
+google-pasta==0.2.0
+gast==0.6.0
+tifffile==2024.9.20
+mdurl==0.1.2
+regex==2024.11.6
+grpcio==1.67.1
+omegaconf==2.3.0
+Werkzeug==3.1.3
+fonttools==4.54.1
+certifi==2024.8.30
+opt_einsum==3.4.0
+keras==3.6.0
+opencv-python-headless==4.10.0.84
+nvidia-ml-py==12.535.161
+matplotlib==3.9.2
+python-dateutil==2.9.0
+scipy==1.14.1
+scikit-image==0.24.0
+jupyter-client==7.3.4
+namex==0.0.8
+Mako==1.3.6
+hydra-core==1.3.2
+wrapt==1.16.0
+comm==0.2.2
+scikit-learn==1.5.2
+imageio==2.36.0
+rich==13.9.4
+transformers==4.46.3
+huggingface-hub==0.26.2
+pandas==2.2.3
+contourpy==1.3.1
+cachetools==5.5.0
+lazy_loader==0.4
+kiwisolver==1.4.7
+cycler==0.12.1
+segmentation_models_pytorch==0.3.5.dev0
+docker-pycreds==0.4.0
+safetensors==0.4.5
+termcolor==2.5.0
+eval_type_backport==0.2.0
+h5py==3.12.1
+tensorflow-io-gcs-filesystem==0.37.1
+tensorboard==2.18.0
+libclang==18.1.1
+jupyterlab_pygments==0.3.0
+webencodings==0.5.1
+colorlog==6.9.0
+defusedxml==0.7.1
+jsonschema-specifications==2024.10.1
+pandocfilters==1.5.1
+efficientnet-pytorch==0.7.1
+jupyter_core==5.7.2
+debugpy==1.6.7
+tornado==6.1
+pydantic==2.9.2
+nbconvert==7.16.4
+bleach==6.2.0
+munch==4.0.0
+tinycss2==1.4.0
+nbclient==0.10.0
+gpustat==1.1.1
+tensorboard-data-server==0.7.2
+markdown-it-py==3.0.0
+jsonschema==4.23.0
+flatbuffers==24.3.25
+nbformat==5.10.4
+smmap==5.0.1
+protobuf==5.28.3
+blessed==1.20.0
+nest_asyncio==1.6.0
+entrypoints==0.4
+nvitop==1.3.2
+tokenizers==0.20.3
+pydantic_core==2.23.4
+alembic==1.14.0
+ipykernel==6.29.5
+tzdata==2024.2
+rpds-py==0.21.0
+antlr4-python3-runtime==4.9.3
+referencing==0.35.1
+albumentations==1.4.18
+absl-py==2.1.0
+annotated-types==0.7.0
+tensorflow==2.18.0
+sentry-sdk==2.18.0
+fastjsonschema==2.20.0
+shapely==2.0.6
+joblib==1.4.2
+threadpoolctl==3.5.0
+timm==0.9.7
